Modifications to build, and run tests on Linux. It has broken CFixedPoint, due to different version of Boost. Need to repair it, and it should be all good.

+++ Source/Interface/Src/DatabaseFixedPoint.inl
@@ -210,7 +210,7 @@ BEGIN_NAMESPACE_DATABASE
 
 				if ( lhs.IsSigned() )
 				{
-					int256_t mul = ( lhs.GetRawValue() * rhr ) / int256_t( GetDecimalMult( lhs.GetDecimals() ) );
+					int256_t mul = ( lhs.GetRawValue() * rhr ) / int256_t( int64_t( GetDecimalMult( lhs.GetDecimals() ) ) );
 
 					if ( rhs.GetDecimals() > lhs.GetDecimals() )
 					{
@@ -228,7 +228,7 @@ BEGIN_NAMESPACE_DATABASE
 				}
 				else
 				{
-					uint256_t mul = ( uint64_t( lhs.GetRawValue() ) * uint256_t( rhr ) ) / uint256_t( GetDecimalMult( lhs.GetDecimals() ) );
+					uint256_t mul = ( uint64_t( lhs.GetRawValue() ) * uint256_t( rhr ) ) / uint256_t( uint64_t( GetDecimalMult( lhs.GetDecimals() ) ) );
 
 					if ( rhs.GetDecimals() > lhs.GetDecimals() )
 					{
@@ -260,11 +260,11 @@ BEGIN_NAMESPACE_DATABASE
 
 				if ( lhs.IsSigned() )
 				{
-					lhr = int256_t( lhr * int256_t( GetDecimalMult( lhs.GetDecimals() ) ) ) / rhr;
+					lhr = int256_t( lhr * int256_t( int64_t( GetDecimalMult( lhs.GetDecimals() ) ) ) ) / rhr;
 				}
 				else
 				{
-					lhr = uint256_t( lhr * uint256_t( GetDecimalMult( lhs.GetDecimals() ) ) ) / uint256_t( rhr );
+					lhr = uint256_t( lhr * uint256_t( uint64_t( GetDecimalMult( lhs.GetDecimals() ) ) ) ) / uint256_t( rhr );
 				}
 
 				return DoShrinkValue( lhr, lhs, rhs );
@@ -277,7 +277,7 @@ BEGIN_NAMESPACE_DATABASE
 
 				if ( rhs.GetDecimals() > lhs.GetDecimals() )
 				{
-					result *= int256_t( GetDecimalMult( rhs.GetDecimals() - lhs.GetDecimals() ) );
+					result *= int256_t( int64_t( GetDecimalMult( rhs.GetDecimals() - lhs.GetDecimals() ) ) );
 				}
 
 				return result;
@@ -289,7 +289,7 @@ BEGIN_NAMESPACE_DATABASE
 
 				if ( rhs.GetDecimals() > lhs.GetDecimals() )
 				{
-					value /= int256_t( GetDecimalMult( rhs.GetDecimals() - lhs.GetDecimals() ) );
+					value /= int256_t( int64_t( GetDecimalMult( rhs.GetDecimals() - lhs.GetDecimals() ) ) );
 				}
 
 				result = int64_t( value );
@@ -306,7 +306,7 @@ BEGIN_NAMESPACE_DATABASE
 			{
 				if ( rhs.GetDecimals() > lhs.GetDecimals() )
 				{
-					value /= uint256_t( GetDecimalMult( rhs.GetDecimals() - lhs.GetDecimals() ) );
+					value /= uint256_t( int64_t( GetDecimalMult( rhs.GetDecimals() - lhs.GetDecimals() ) ) );
 				}
 
 				uint64_t tmp = uint64_t( value );
@@ -357,7 +357,7 @@ BEGIN_NAMESPACE_DATABASE
 
 	inline int64_t CFixedPoint::ToInt64()const
 	{
-		return int64_t( _value / int256_t( GetDecimalMult( _decimals ) ) );
+		return int64_t( _value / int256_t( int64_t( GetDecimalMult( _decimals ) ) ) );
 	}
 
 	inline uint64_t CFixedPoint::ToUInt64()const
