+++ Source/Test/Src/DatabaseFixedPointTest.cpp
@@ -22,7 +22,7 @@ BEGIN_NAMESPACE_DATABASE_TEST
 		void CheckFixedPoint( uint8_t precision, uint8_t decimals, int64_t value, const String & sval )
 		{
 			double dval = value / pow( 10, decimals );
-			int64_t ival = int64_t( value / int256_t( pow( 10, decimals ) ) );
+			int64_t ival = int64_t( value / int256_t( uint64_t( pow( 10, decimals ) ) ) );
 			CFixedPoint fp( value, precision, decimals );
 			BOOST_CHECK_EQUAL( fp.GetRawValue(), value );
 			BOOST_CHECK_EQUAL( fp.ToDouble(), dval );
@@ -93,13 +93,13 @@ BEGIN_NAMESPACE_DATABASE_TEST
 		int decimals = 0;
 		int64_t value;
 		CLogger::LogInfo( StringStream() << "  Max Value" );
-		value = int64_t( std::numeric_limits< int64_t >::max() % int256_t( pow( 10, precision ) ) );
+		value = int64_t( std::numeric_limits< int64_t >::max() % int256_t( uint64_t( pow( 10, precision ) ) ) );
 		CheckFixedPoint( precision, decimals, value, STR( "0" ) );
 		CLogger::LogInfo( StringStream() << "  Min Value" );
-		value = int64_t( std::numeric_limits< int64_t >::min() % int256_t( pow( 10, precision ) ) );
+		value = int64_t( std::numeric_limits< int64_t >::min() % int256_t( uint64_t( pow( 10, precision ) ) ) );
 		CheckFixedPoint( precision, decimals, value, STR( "0" ) );
 		CLogger::LogInfo( StringStream() << "  Lowest Value" );
-		value = int64_t( std::numeric_limits< int64_t >::lowest() % int256_t( pow( 10, precision ) ) );
+		value = int64_t( std::numeric_limits< int64_t >::lowest() % int256_t( uint64_t( pow( 10, precision ) ) ) );
 		CheckFixedPoint( precision, decimals, value, STR( "0" ) );
 		CLogger::LogInfo( StringStream() << "  Invalid Value (greater precision than allowed)" );
 		BOOST_CHECK_THROW( CFixedPoint( std::numeric_limits< int64_t >::max(), precision, decimals ), CDatabaseException );
@@ -114,13 +114,13 @@ BEGIN_NAMESPACE_DATABASE_TEST
 		int decimals = 0;
 		int64_t value;
 		CLogger::LogInfo( StringStream() << "  Max Value" );
-		value = int64_t( std::numeric_limits< int64_t >::max() % int256_t( pow( 10, precision ) ) );
+		value = int64_t( std::numeric_limits< int64_t >::max() % int256_t( uint64_t( pow( 10, precision ) ) ) );
 		CheckFixedPoint( precision, decimals, value, StringUtils::ToString( value ) );
 		CLogger::LogInfo( StringStream() << "  Min Value" );
-		value = int64_t( std::numeric_limits< int64_t >::min() % int256_t( pow( 10, precision ) ) );
+		value = int64_t( std::numeric_limits< int64_t >::min() % int256_t( uint64_t( pow( 10, precision ) ) ) );
 		CheckFixedPoint( precision, decimals, value, StringUtils::ToString( value ) );
 		CLogger::LogInfo( StringStream() << "  Lowest Value" );
-		value = int64_t( std::numeric_limits< int64_t >::lowest() % int256_t( pow( 10, precision ) ) );
+		value = int64_t( std::numeric_limits< int64_t >::lowest() % int256_t( uint64_t( pow( 10, precision ) ) ) );
 		CheckFixedPoint( precision, decimals, value, StringUtils::ToString( value ) );
 
 		CLogger::LogInfo( StringStream() << "**** End TestCase_FixedPointMaxPrecisionMinDecimals ****" );
@@ -135,15 +135,15 @@ BEGIN_NAMESPACE_DATABASE_TEST
 		int64_t value;
 		String sval;
 		CLogger::LogInfo( StringStream() << "  Max Value" );
-		value = int64_t( std::numeric_limits< int64_t >::max() % int256_t( pow( 10, precision ) ) );
+		value = int64_t( std::numeric_limits< int64_t >::max() % int256_t( uint64_t( pow( 10, precision ) ) ) );
 		sval = "0." + StringUtils::ToString( value );
 		CheckFixedPoint( precision, decimals, value, sval );
 		CLogger::LogInfo( StringStream() << "  Min Value" );
-		value = int64_t( std::numeric_limits< int64_t >::min() % int256_t( pow( 10, precision ) ) );
+		value = int64_t( std::numeric_limits< int64_t >::min() % int256_t( uint64_t( pow( 10, precision ) ) ) );
 		sval = "-0." + StringUtils::ToString( uint64_t( abs( value ) ) );
 		CheckFixedPoint( precision, decimals, value, sval );
 		CLogger::LogInfo( StringStream() << "  Lowest Value" );
-		value = int64_t( std::numeric_limits< int64_t >::lowest() % int256_t( pow( 10, precision ) ) );
+		value = int64_t( std::numeric_limits< int64_t >::lowest() % int256_t( uint64_t( pow( 10, precision ) ) ) );
 		sval = "-0." + StringUtils::ToString( uint64_t( abs( value ) ) );
 		CheckFixedPoint( precision, decimals, value, sval );
 
@@ -159,17 +159,17 @@ BEGIN_NAMESPACE_DATABASE_TEST
 		int64_t value;
 		String sval;
 		CLogger::LogInfo( StringStream() << "  Max Value" );
-		value = int64_t( std::numeric_limits< int64_t >::max() % int256_t( pow( 10, precision ) ) );
+		value = int64_t( std::numeric_limits< int64_t >::max() % int256_t( uint64_t( pow( 10, precision ) ) ) );
 		sval = StringUtils::ToString( value );
 		sval.insert( ( sval.rbegin() + decimals ).base(), STR( '.' ) );
 		CheckFixedPoint( precision, decimals, value, sval );
 		CLogger::LogInfo( StringStream() << "  Min Value" );
-		value = int64_t( std::numeric_limits< int64_t >::min() % int256_t( pow( 10, precision ) ) );
+		value = int64_t( std::numeric_limits< int64_t >::min() % int256_t( uint64_t( pow( 10, precision ) ) ) );
 		sval = StringUtils::ToString( value );
 		sval.insert( ( sval.rbegin() + decimals ).base(), STR( '.' ) );
 		CheckFixedPoint( precision, decimals, value, sval );
 		CLogger::LogInfo( StringStream() << "  Lowest Value" );
-		value = int64_t( std::numeric_limits< int64_t >::lowest() % int256_t( pow( 10, precision ) ) );
+		value = int64_t( std::numeric_limits< int64_t >::lowest() % int256_t( uint64_t( pow( 10, precision ) ) ) );
 		sval = StringUtils::ToString( value );
 		sval.insert( ( sval.rbegin() + decimals ).base(), STR( '.' ) );
 		CheckFixedPoint( precision, decimals, value, sval );
